Delete test db only

+++ androidTest/java/com/gbrfix/randomyzik/PlaylistDbTest.java
@@ -30,17 +30,12 @@ public class PlaylistDbTest {
     public void setUp() throws Exception {
         c = InstrumentationRegistry.getInstrumentation().getTargetContext();
         dbService = new DbService(c);
-        //TODO: Rechercher toutes les base donnees test et les vider
         String[] dbNames = c.databaseList();
         for (String dbName : dbNames) {
-            c.deleteDatabase(dbName);
+            if (dbName.startWith("test-")) {
+                c.deleteDatabase(dbName);
+            }
         }
-        /*MediaDAO dao = new MediaDAO(c, MediaDAOTest.TEST_DBNAME);
-        dao.open();
-        dao.getDb().delete("medias", null, null);
-        SQLiteCursor cursor = dao.getAll();
-        assertEquals(0, cursor.getCount());
-        dao.close();*/
     }
 
     @Test
